Created generic `reportError` function
