+++ packages/frontend/web/components/MostViewed/MostViewed.tsx
@@ -13,7 +13,7 @@ import { Container } from '@guardian/guui';
 import { namedAdSlotParameters } from '@frontend/model/advertisement';
 import { AdSlot, labelStyles } from '@frontend/web/components/AdSlot';
 import { OutbrainContainer } from '@frontend/web/components/Outbrain';
-import * as Sentry from '@sentry/browser';
+import { reportError } from '@frontend/web/browser/sentry/sentry';
 import { useApi } from '@frontend/web/components/lib/api';
 
 import { MostViewedGrid } from './MostViewedGrid';
@@ -128,10 +128,7 @@ export const MostViewed = ({ config, sectionName, pillar }: Props) => {
     const { data, error } = useApi<TabType[]>(url);
 
     if (error) {
-        Sentry.withScope(() => {
-            Sentry.setTag('feature', 'most-viewed');
-            Sentry.captureException(error);
-        });
+        reportError(error, 'most-viewed');
     }
 
     if (data) {
