+++ packages/frontend/web/components/ShareCount.tsx
@@ -5,7 +5,7 @@ import ShareIcon from '@guardian/pasteup/icons/share.svg';
 import { textSans } from '@guardian/pasteup/typography';
 import { from, wide, leftCol } from '@guardian/pasteup/breakpoints';
 import { integerCommas } from '@frontend/lib/formatters';
-import * as Sentry from '@sentry/browser';
+import { reportError } from '@frontend/web/browser/sentry/sentry';
 
 const shareCount = css`
     ${textSans(6)};
@@ -88,10 +88,7 @@ export class ShareCount extends Component<Props, { shareCount?: number }> {
                 });
             })
             .catch(err => {
-                Sentry.withScope(() => {
-                    Sentry.setTag('feature', 'share-count');
-                    Sentry.captureException(err);
-                });
+                reportError(err, 'share-count');
             });
     }
